requirements.txt
flask
boto3
python-dotenv
gunicorn
flask-limiter

# Development dependencies
pytest 